move windows specific options to its own file
